--- OhMyRunPod/modules/pod_info/__init__.py
@@ -1,2 +1,0 @@
-# OhMyRunPod/modules/pod_info/__init__.py
-from .pod_info import print_pod_info
