+++ OhMyRunpod/modules/pod_info/__init__.py
@@ -1,0 +1,2 @@
+# OhMyRunpod/modules/pod_info/__init__.py
+from .pod_info import print_pod_info
